- adaption to new solver - reformati

+++ ttb_worldmodel/src/ttb/wm/topology/AnnotatedGrid.cpp
@@ -7,31 +7,27 @@ namespace ttb
 namespace wm
 {
 
-AnnotatedGrid::AnnotatedGrid(ttb::TTBWorldModel *wm)
-    : wm(wm)
+AnnotatedGrid::AnnotatedGrid(ttb::TTBWorldModel* wm)
+        : wm(wm)
 {
     auto sc = supplementary::SystemConfig::getInstance();
     std::string annotatedGridTopic = (*sc)["TTBWorldModel"]->get<std::string>("AnnotatedGrid.annotatedGridTopic", NULL);
     std::string gridTopic = (*sc)["TTBWorldModel"]->get<std::string>("AnnotatedGrid.gridTopic", NULL);
     ros::NodeHandle n;
-    this->annotatedGridSubscriber =
-        n.subscribe(annotatedGridTopic, 10, &AnnotatedGrid::receiveAnnotatedGridPoints, (AnnotatedGrid *)this);
+    this->annotatedGridSubscriber = n.subscribe(annotatedGridTopic, 10, &AnnotatedGrid::receiveAnnotatedGridPoints, (AnnotatedGrid*) this);
     this->gridPublisher = n.advertise<ttb_msgs::Grid>(gridTopic, 5, false);
 
     this->sendGridPoints();
 }
 
-AnnotatedGrid::~AnnotatedGrid()
-{
-}
+AnnotatedGrid::~AnnotatedGrid() {}
 
 void AnnotatedGrid::sendGridPoints()
 {
     auto gridPoints = this->generateGridPoints();
 
     ttb_msgs::Grid grid;
-    for (auto point : gridPoints)
-    {
+    for (auto point : gridPoints) {
         grid.points.push_back(point);
     }
     this->gridPublisher.publish(grid);
@@ -54,23 +50,18 @@ std::vector<geometry::CNPointAllo> AnnotatedGrid::generateGridPoints()
 void AnnotatedGrid::receiveAnnotatedGridPoints(ttb_msgs::AnnotatedGridPtr annotatedGrid)
 {
     this->gridMap.clear();
-    for (int i = 0; i < annotatedGrid->points.size(); i++)
-    {
+    for (size_t i = 0; i < annotatedGrid->points.size(); i++) {
         auto poi = this->wm->topologicalModel.getPOIByName(annotatedGrid->annotatedPOIs[i]);
 
-        if (!poi)
-        {
+        if (!poi) {
             std::cerr << "AnnotatedGrid: Received poi " << annotatedGrid->annotatedPOIs[i] << " is unknown!" << std::endl;
             continue;
         }
 
         auto mapEntry = this->gridMap.find(poi->room);
-        if (mapEntry != this->gridMap.end())
-        {
+        if (mapEntry != this->gridMap.end()) {
             mapEntry->second.push_back(AnnotatedGridPoint(poi->room, annotatedGrid->points[i]));
-        }
-        else
-        {
+        } else {
             this->gridMap.emplace(poi->room, std::vector<ttb::wm::AnnotatedGridPoint>());
             mapEntry->second.push_back(AnnotatedGridPoint(poi->room, annotatedGrid->points[i]));
         }
@@ -80,12 +71,9 @@ void AnnotatedGrid::receiveAnnotatedGridPoints(ttb_msgs::AnnotatedGridPtr annota
 std::vector<ttb::wm::AnnotatedGridPoint> AnnotatedGrid::getGridOfRoom(std::shared_ptr<ttb::wm::Room> room)
 {
     auto mapEntry = this->gridMap.find(room);
-    if (mapEntry != this->gridMap.end())
-    {
+    if (mapEntry != this->gridMap.end()) {
         return mapEntry->second;
-    }
-    else
-    {
+    } else {
         std::cerr << "AnnotatedGrid: Requested room " << room->name << " is unknown!" << std::endl;
         return std::vector<ttb::wm::AnnotatedGridPoint>();
     }
